pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "nhl_score_prediction"
authors = [
	{name = "Brent Barbachem", email = "barbacbd@gmail.com"},
]
description = "NHL package/executable to predict NHL scores."
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
keywords = [
    "python",
    "machine-learning",
    "statistics",
    "sports",
    "prediction",
    "python3",
    "pytorch",
    "predictive-modeling",
    "nhl-core",
    "sports-stats",
    "sports-analytics",
    "event-analysis"
]
classifiers = [
    "Intended Audience :: Science/Research",
    "License :: Freeware",
    "Natural Language :: English",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development"
]
dependencies = [
    "inquirer",
    "scipy",
    "keras",
    "tensorflow",
    "numpy",
    "pandas",
    "openpyxl",
    "mrmr_selection",
]
dynamic = ["version"]

[tool.setuptools]
include-package-data = true

[project.optional-dependencies]
tests = [
    "pytest",
    "coverage",
    "pytest-cov",
    "mock",
    "pylint",
    "pytest-order"
]
clean = [
    "pyclean",
]
[project.scripts]
nhl-poisson-predict = "nhl_score_prediction.poisson.poisson:main"

[tool.setuptools.package-data]
'src' = ['nhl_score_prediction/support/schedules/2022/*.json']


